fix get backend address on host
